Quite a few changes but mostly a bit naff
I'm trying to move the server code about so that it's a runnable object,
this way I can put it in my main game loop (or something) and set it
running for the duration of the game program. I've also added somethign
called message_queu the idea being that the server puts the messages in
that queue and some other bit of code pulls them all out and processes
them. Currently all the messages are pulled otu at once which I think is
ok because they can just ben put into another processing buffer
somewhere else until they're dealt with properly.